pyproject.toml
[project]
name = "slide-agents"
version = "0.1.0"
description = "Flask app with Google OAuth (People API) for slideAgents"
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "dnspython>=2.7.0",
    "flask>=3.1.2",
    "google-api-python-client>=2.181.0",
    "google-auth-oauthlib>=1.2.2",
    "mongomock>=4.3.0",
    "pymongo>=4.14.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "langgraph>=0.2.40",
    "langchain>=0.3.7",
    "langchain-google-genai>=2.0.5",
    "langsmith>=0.2.3",
    "pydantic>=2.8.2",
]

[tool.uv]
# Allow uv to resolve and manage dependencies.
